app/requirements.txt
backports.ssl-match-hostname
flask
beautifulsoup4
peewee
psycopg2-binary
pyowm
python-dateutil
PyYAML
requests
six
slackbot
slacker
websocket-client
